feat: Task 4 — byte-exact Unigram conformance vs transformers (t5-small)
- float→double for scores/best in unigram.h, unigram.c, ug_class.c: matches
  HF tokenizers library (f64) and avoids precision-induced Viterbi flip.
- Fix unk_score computation: use min(non-zero scores) − 10.0 (matches
  tokenizers Rust impl); t5-small stores unk/pad/eos with score 0.0 so the
  old scores[unk_id] path would treat unk as always-preferred.
- Encoding.php: add prepend_scheme ('always'/'first'/'never') mapping so
  t5-small's pre_tokenizer JSON is read correctly.
- Generator: add Unigram section — T5TokenizerFast, 24 curated ASCII cases,
  emits t5_tokenizer.json + unigram_conformance.json (committed).
- New test 032-unigram-conformance.phpt: ALL CONFORMANT (24 cases, t5-small).
- Full suite 14/14 pass.

+++ src/unigram.c
@@ -6,7 +6,7 @@
 static const char MS[3]={(char)0xE2,(char)0x96,(char)0x81}; /* U+2581 */
 static int boundary(const char *s,size_t i){ return (s[i]&0xC0)!=0x80; } /* codepoint start */
 
-int tk_unigram_encode(const tk_model *m,const float *scores,const char *text,size_t len,
+int tk_unigram_encode(const tk_model *m,const double *scores,const char *text,size_t len,
                       const tk_ug_opts *o,uint32_t **out_ids,size_t *n_out){
     /* Metaspace: prefix + replace spaces with ▁ */
     size_t cap=len*3+3; char *buf=malloc(cap?cap:1); size_t blen=0;
@@ -16,14 +16,14 @@ int tk_unigram_encode(const tk_model *m,const float *scores,const char *text,siz
         else buf[blen++]=text[i];
     }
     size_t n=blen;
-    /* Viterbi */
-    float *best=malloc((n+1)*sizeof(float));
+    /* Viterbi — use double precision to match the HF tokenizers library (f64). */
+    double *best=malloc((n+1)*sizeof(double));
     size_t *bp_i=malloc((n+1)*sizeof(size_t));
     uint32_t *bp_id=malloc((n+1)*sizeof(uint32_t));
-    for(size_t j=0;j<=n;j++){ best[j]=-FLT_MAX; bp_i[j]=(size_t)-1; bp_id[j]=o->unk_id; }
+    for(size_t j=0;j<=n;j++){ best[j]=-DBL_MAX; bp_i[j]=(size_t)-1; bp_id[j]=o->unk_id; }
     best[0]=0;
     for(size_t i=0;i<n;i++){
-        if(best[i]==-FLT_MAX) continue;
+        if(best[i]==-DBL_MAX) continue;
         if(!boundary(buf,i)) continue;
         int matched=0;
         size_t jmax = i+o->max_piece_len; if(jmax>n) jmax=n;
@@ -31,7 +31,7 @@ int tk_unigram_encode(const tk_model *m,const float *scores,const char *text,siz
             if(j<n && !boundary(buf,j)) continue; /* only at codepoint boundaries */
             uint32_t id=tk_model_rank(m,(const uint8_t*)(buf+i),j-i);
             if(id==TK_RANK_MAX) continue;
-            float s=best[i]+scores[id];
+            double s=best[i]+scores[id];
             if(s>best[j]){ best[j]=s; bp_i[j]=i; bp_id[j]=id; matched=1; }
         }
         /* unk fallback: one codepoint */
